[##] Refacto : Removed siretValidator class
